updateme: auto-generate board definitions
When we are bootstrapping a BSP, the basic information to create
the branching and configuration can be generated by the system
itself.

In this case, we create a basic .scc file that includes the
appropriate base branch and define the requirement arch/branch/board
information.

We also look for defconfig's, since during bootstrap they are
allowed and commonly used.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>