updateme: only search .scc files for duplicate configs
To prevent a double include of a 'floating' config fragment, the
source directory of the fragment is checked for an explicit
include by a feature (a .scc file). But the test was grepping
all files in the source directory. Which meant that defconfig,
that contained a defconfig could actually exclude itself.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>