createme: improve BSP bootstrapping
To streamline BSP bootstrapping, createme now actually
creates the required branch rather than only dumping it
in a .scc file. Later phases of the build will create
the .scc file, so it isn't required here.

Also, a BSP with a '-' in the name would break the parent
branch detection. Improving the parsing of this allows more
flexibile branch naming.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>