configme: place in-tree scripts at the end of the path
configme can be used outside of a build system, so it looks
for some of the scripts it uses within the tree being configured
itself and places them on the path.

But when running within a build system, we should prefer the
ones from the build system and not the in-tree variants. To
fix this, we add the in tree directory to the END of the path
to allow build system ones to take precedence.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>