kgit-checkpoint: query branches for checkpoint directories
The test for the type of checkpoint meta directory was
filesystem based. But this check won't work before you
uncheckpoint a tree (since the directory won't be in the
working tree of all branches).

To fix this, we check the checkpoint branch itself to see
what the meta directory is called.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>