configme: take the output directory as an option
configme used to be able to calculate the output/build directory
when branches were always <machine>-<kernel type>. Branch names
can now be widely different and to avoid embedding complexity
in the scripts it is easier to just pass ${B} from the build system
down to the scripts. We enable this by adding a --output option
to the script.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>