kconf_check: strip /base from branches
A suffix of /base on the end of a branch is interpreted by
the system and is only required due to on-disk git implementation
details. So if we get a branch that ends in /base, we should
be stripping it off since we won't find config fragements
in that directory.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>