kgit-meta: recreate missing branches
If a repository has a single branch, but a meta directory that
describes more branches, we have all the information required
to validate the repo. In order to validate safely, and not
create special cases, re-creating the branches during the
processing is done.

The subroutines responsible for picking branch names, and
renaming existing ones are improved to detect that branches
are being created, and to create them safely.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>