configme: determine meta branch based on directories, not branch naming
previously the meta directory was being deterined based on the branch
naming structure. If it had a / it must be a 'meta' directory. If it
didn't have one, it must be a 'wrs' format. This doesn't hold in the
general case, so it is better to simply test for the 'meta' subdir
and tag things appropriately.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>