updateme: allow the location of board descriptions based on defines
The existing infrastruce used scc_leaf as the method for locating
a desired board description. This meant that the branching information
had to be in a very particular format to be recognized as something
we could process and build. Allowing a board description to be
recognized by its defines allows us to locate and build multiple
board descriptions for a single branch.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>