updateme: migration board definitions into kernel tree
To allow BSP definitions to exist outside the tree, and then
migrate into the kernel source tree .. making them available for
processing as a BSP definition, special handling is required.

If a .scc file is detected that defines a leaf node, it is NOT
added to the end of the feature list automatically, but is instead
copied into the tree for potential location by the leaf node
logic.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>