generate_cfg: allow redefinition of non-hardware to hardware
Previously once an option was declared as non-hardware, it was
set it stone. Any overrides/changes of non-hardware values result
in a configuration check warning, which means there is no way
to silence these warnings if you know what you are doing.

To fix this, if an option is overridden and an accompanying
hardware.CF declaration is provided, we won't classify it as
a redefinition warning. Instead it will be logged as an option
that transitioned to board specific. This allows auditing of the
conifer's, but prevents warnings that can't be inhibited.

Signed-off-by: Bruce Ash field <bruce.ashfield@windriver.com>