kgit-meta: add error checking and autoconfiguration to git commands
Fixes [YOCTO #1838]

When git commands are embedded into feature descriptions, they can
fail for a variety of reasons. The current meta series processing does
not trap errors, and will continue to process potentially failing
commands. All git commands are now checked for their return code, and
in case of error, the error code is passed up via early exit.

A second common git issue is on a new machine that has not been
configured with a username/email for commits. This is now detected
by checking for a global user.name. If one is not found, a local
user name of "Auto Configured" and email of "auto.configured" is
created to allow git operations to succeed.

Signed-off-by: Bruce Ashfield <bruce@windriver.com>