kern-tools: remove branchname assumptions
Multiple tools process the meta series to determine the final
branch name that a given series would create. This means that
tweaking in branch naming logic ripples through many tools and
scripts. Unifying the branch logic in kgit-meta, and providing
the ability to simply execute branching commands allows other
tools to assume that the current branch is always the correct
branch to process.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>