kern-tools: add required and optional kernel configuration processing
In addition to the hardware/non-hardware classification of kernel config
options, with this commit is it possible to specify 'required' and
'optional' kernel configuration values in feature descriptions. This allows
kernel configuration policy to be more than just BSP and "non-BSP".

If required options are missing, or defined and changed a WARNING during
the kernel config validation phase will be displayed. This warning can
be optionally turned into a hard error.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>