checkpoint: allow orphan branches and drop unused functionality
The ability to 'join' checkpoints is not used, and overly complicates
the checkpoint processing, so it can be removed.

Secondarily the checkpoint creation and restore is changed to handle
branches that do not share a common base with the rest of the repository.
With this change orphan branches or common branches can be used to
store meta data.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>