kgit-init: remove remote init support and update default content branch
The ability to run kgit-init against a remote repository is unused,
and it is reasonable to expect that a build system / fetcher should
be used to ensure that the repository being used is local the location
where it will be built. As a result, we can remove support from the
script.

Secondarily, a safe default content branch of 'master' is chosen, since
it will exist in all repositories.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>