createme: fix checkpoint restoration for reset branches
The meta branch can optionally be merged out to BSP branches. This removes
the need to restore the checkpoint when working with the tree.  The way
it detects the merge is by checking to see how many branches contain the
meta data. If there's more than one, the branch was was merged out.

Unless you are a BSP that isn't tracking the latest meta, and you get
meta and meta-orig created. That's two branches and the code opts to not
restore the checkpoint, which leads to configuration errors.

The fix is simple. We allow for 2 or less branches with meta, and will
still restore the checkpoint. Three and up, we won't.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>