kern-tools: remove unused code, scripts and transition code
The period of supporting old trees with a different meta
branch name and directory structure are gone. So the cleanup
and removal of the old structure can be completed.

The meta branch and directory are now controlled via command line,
or via the KMETA environment variable. No testing and conditional
processing of the tree are required.

Additionally, the generate_cfg script is no longer used, or is the
branch conditing code in createme. So they can be safely removed
from the tools and repository.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>