kgit-publish: remove --remote option
The ability to publish and automatically push a repository was
never used, and is error prone. The complexit isn't needed in
the script, so removing it is the best option.

An explicit push after tree publication is suggested, or a
wrapper script (specific to a particular infrastructure) around
this script.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>