kgit-scc: add text for no arg and invalid arg case.
The value here is when you use "-j30" vs. "-j 30".  The former
will silently fail since this isn't true getopt and you get
a non deterministic usage message.  But if you get a message:

	Invalid arg "-j30".

as well as the usage message, then chances are you'll clue in andi
be able to suss things out from there.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>