buildall: script to independently build all board kernels
For sanity testing, it is nice to be able to build all the
possible kernels with nothing more than the kernel git repo
and a handful of cross compilers.

This script achieves that by determining what all the possible
boards are, what their respective branch is, and then configuring
and compiling each and every one of them.

A sample RC file can be copied to $HOME/.buildallrc and then
customized to suit local variations -- e.g. paths to toolchains,
path to the tree that you want to test, and so on.

The defaults assume you've already got a populated sysroot from
building the main architectures at least once already in order
to provide toolchains, but there is no requirement to use the
yocto generated toolchains from the sysroot.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>