buildall: support semi seamless restarts
In a case where buildall finds one failure in a larger set, you
may not want to re-run all the passed builds.  To that end, being
able to "resume" in an existing buildall space, by retrying the
builds which failed (and skipping those which passed), can be very
handy for finalizing a tree.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>