configme: relieve it of all knowledge of scc files
Everything configme needs is contained in the meta-series.
Having it go after multiple sources of information can just
lead to confusion, esp. if the meta-series is out of sync
with the scc file specified in meta/top_tgt.

Make configme use the meta series as the authoratative source
of information, and strip out all magic around trying to find
and rate/rank scc files.

This doesn't make use of the added KMACHINE setting that is
parsed out, but it will be used in the future to make the
command line value of $machine redundant/obsolete.

Also make the regex matches on the key defines more strict.
While the scc files can have comments, the meta-series does
not, so in theory we shouldn't need to anchor the regex to BOL
(but do so anyway, just to be 100% safe).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>