buildall: switch back to scc driven processing
This is a partial revert of commit 5b6498c4bc02a

    buildall: remove all instances of it using/reading scc files

The above was done, since the final branch used by an scc
file is not obvious (imagine scc A that includes scc B which
in turn does the actual branch selection).  Since the output
of scc (the meta-series) clearly shows the final branch, the
switch over seemed obvious.

In the context of building all BSPs from a freshly generated
tree, we know the meta series files match what is on all the
branches.  However if someone wants to use buildall on a
tree that is in maintenance mode, there may have been new
scc added (i.e. no meta), or configuration fragments relocated
(meta with stale paths) resulting in odd failures that would
not happen if the branch was built via the complete build
system.

This puts things back to being driven from the scc leaf nodes.
The stale meta problem is handled by regenerating them.  There
is an optional "use what is there" knob that can be used by
people generating a fresh tree, since they know that there is
no risk of staleness in that one particular use case (and since
regeneration of them can be timewise expensive).

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>