scc/kgit-meta: detect and avoid duplicating patching
It is common that a feature fragment is included multiple times,
and these double inclusions can cause problems if a fragment includes
both patches and configuration.

This was previously dealt with by splitting the feature into an
'enable' configuration fragment and a 'feature' fragment that applies
the patches. This is effective, but requires that knowledge of the
entire include structure be known when creating a new feature.

We can detect when a patch has already been applied, or when a
feature fragment is included multiple times and do two things:

  - emit a clear message about the multiple includes
  - skip the double application of the patches via tracking a global
    series file

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>