updateme/scc: allow config fragment exclusion
It is common to need the features (patches, git operations) of a
branch, but not want the kernel configuration fragments of a given
branch. To allow this, we provide a new include flag "nocfg".

When this flag is used, all of the configuration fragments included
by the targetted feature will not be applied to the current build,
with one exception, a base/critical fragment can force it's config
values, since without them, the system would not be functional.

Example:

   include ktypes/standard/standard.scc nocfg

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>