scc: wrap git merge
To allow for parameter validation and sanity checking, wrap "git merge"
as a dedicated "merge" command instead of using the raw git fallback.

This also makes it consistent with existing top level commands such
as  'tag', 'branch', 'patch', etc.

There are no changes to arguments, and existing 'git merge' commands
continue to work with this change.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>