kconf_check: fix new Kconfig detection
One of the functions of the kernel configuration audit is to notify
the user if Kconfig* files have been removed from the kernel, and
also to notify of new Kconfig files.

New Kconfig files should be classified as hardware or non-hardware to
allow BSP audits to notify if boards are setting values that they
shouldn't, hence why notifying about new "buckets" is important.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>