scc/updateme/spp: fix cfg includes and remove obsolete code
To support inhibiting of configuration values via include directives,
it must happen in a depth first based processing mode. The existing
monolithic processing of files, which then generate a script, that
generates a meta-series work in a model where all files are found,
individually processed and then re-assembled to generate the file output.

To allow enhanced processing of includes, we break the processing into
two parts, while keeping the output identical to the existing outputs
of the scripts. The first phase preprocesses the target files, in a
depth first manner, expanding included files, and converting kernel
configuration and patch directives to absolute paths (while performing
checks on their existence). The second phase, consumes the preprocessed
file, and generates a meta-series for future processing.

The logic and output of the existing scc processing is maintained, but
is now faster, robust and can handle more complex conditioning of the
input files.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>