scc: allow patch and configuration searching
A regression was reported that unless a patch file was local to the .scc
file, it could no longer be found. This is due to the searching for
patch/config files not taking into account a locally passed set of include
paths versus the global ones set via -I <path>

This error could have been masked due to an exit code not propagating
to the final exit code of 'spp', so we fix that as well in this change.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>