updateme: use absolute path for generated BSP descriptions
When a custom BSP is used, a top level BSP is generated by the tools and fed
to the build system just as a user defined BSP would be located and
passed. The location of the generated file is placed in the top_tgt file,
which is used by subsequent stages. A relative path was being placed into
top_tgt, which binds the build to a particular directory structure and
working directory.

The location of parts of the build have changed, and this relative path is
no longer accurate. Changing it to an absolute path solve the build issues
related to custom BSPs.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>