updateme: expand features that point to a directory name
When a feature is passed to the kernel configuration scripts, and that
feature is a directory name, it is a shortcut for:

  $DIR/$DIR.scc

This expansion is not commonly used, and should be avoided. But for the
purposes of backwards compatibility, updateme can expand the feature into
a .scc file before passing it to the next set of configuration scripts.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>