updateme: enforce a matching machine
When searching for a matching .scc file to a KDESC. We pick the
highest scoring .scc file. But this isn't useful if we don't
match the KMACHINE. We can test the 'winner' for a maching board
and clear the selection if we don't match, which fixes the problem
of unexpected boards being selected.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>