kgit: remove guilt dependency
The kern-tools use guilt as a patch management system, but with the
ability to merge branches, and do iterative development most of the
guilt safe guards need to be disabled. The resulting complexity to
both guilt and the kern-tools outweighs the benifits.

This change simplifies kgit-meta and introduces a new script called
kgit-s2q (series-to-queue), which manages the series file that was
previously processed by guilt.

This script can deal with autoresume, validates patches against
the tree (instead of against a status file), can deal with merges
and locally applied changes.

No user visible functionality is impacted by this change.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>