kgit: speed patch application by batching patches
kgit-meta knows if it just added a patch to the series file, so we
can pass it to kgit-s2q as the starting point. This avoid auto resume
detection and speeds patch application.

Also as part of this change, the variable naming is cleaned up to
differentiate between the resume fencepost detection, and the command
line option, which is the exact patch to start with.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>