kern-tools: use .meta as meta data container
Previously the meta directory was added to all .gitignore files, otherwise
all users of the kernel would see an untracked directory "meta". This
caused workflow issues, since the same .gitignore was shared on the meta
branch itself. Which meant that changes were not shown and files needed to
be added with the -f flag.

To fix this, we teach the tools to look for .$META as the meta data
container, and place the meta branches directory under that same name.

The result is that the .meta directory is ignored by git automatically, but
is available to all branches. And switching to the meta branch no longer
warns of clobbered files, and once you arrive, files are tracked and
manipulated normally.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>