kgit-s2q: Restore implicit exit status to "git apply" section
The code taken from "git quiltimport" used to have && at the
end of each line to give a staged execution of steps.  The
addition of having the "git am" metadata created in rebase-apply
inadvertently removed them, causing the steps to be run
unconditionally.  Restore the old behaviour, or we'll get into
trouble if we mix "git am" applications with "git apply" ones.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>