kgit-s2q: add proper commit ID handling for mixed am/apply usage
The code taken from quiltimport unconditionally used "git apply"
and hence "commit" only needed to be initialized once and then
it would be constantly updated in the loop.  But we want to
support mixed use of "git am" and "git apply" so we need to ensure
the "commit" stays up to date if/when "git am" is used.  Otherwise
commit-tree will essentially bundle all am'd patches in with the
current "apply" operation, which is not good.

Update the value after we've used "git am" to keep things in sync.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>