kgit-s2q: make Applying messages consistent.
If we run "git am" then we rely on the "Applying: " message to
come from that.  But it isn't conditional on "verbose" flag.
When we use "git apply" the message is conditional on verbose.

Clobber the message in both cases, and simply echo out the patch
name, which is a more valuable piece of information anyways -- esp.
if a fixup of it is required.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>