kgit-s2q: make a "git am" failure a soft error
Previously, if git-am failed, it was a hard error and
processing would immediately halt.  Replace that with
a soft-error INFO message, and let it fall through to
the manual commit manipulation/generation.  Since the
"--gen" flag is normally set, the ordering would be
"apply" 1st and "am" 2nd, so this is a no-op at this time.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>