kgit-s2q: default to using "git am" 1st.
The fast path through, and the only reliable way to know if
"git am" will succeed, is to simply try it (since "git apply --check"
is not a guarantee git-am will work).

So do that 1st.  If it fails, abort the git-am, and let the fallback
solution be to manually craft up a commit and allow fuzz.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>