kgit-s2q: Try reduced context with "git am" before falling back to "git apply"
Most patches in a tree generation context have proper headers and so on,
and hence do not need the whole "git apply" mechanism with the manual
creation of subject/date/author.  They will apply as-is with just some fuzz.
So, try "git am" with reduced context before falling back to "apply" mode.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>