kgit-s2q: log patches that don't pass "git am"
If the patch needs to be processed with reduced context, or it
needs manual application with "git apply" then log its name and
the associated new commit ID in a refresh file.

This will facilitate easy inspection of what needs a refresh and
if the auto refresh during treegen looks correct.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>