kgit-s2q: locate top of series via subjects
If the matching of the top commit to the series fails, we flip processing
and look for the top of the series in the current branch. That check is done
via the git patch-id of the top patch comparies to the rev-list of the
branch.

If a patch has been applied with fuzz, this approach fails. So much the same
way that we compare the subect of the top commit against the series, we can
search for the top patch in the series against the subjects in the rev-list.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>