kgit-s2q: support no author patches
previous versions of the kern-tools supported the ability to import a bare
patch, with no From: Subject: or other identifying fields that are typically
in a full commit.

The same type of commit with kgit-s2q will prompt for a author ID, just
as git-quilt-import does. In build system environment that leads to an
infinite loop and the commit is never pushed.

To fix this issue, we add an interactive flag (-i), that when passed the
prompt based behaviour is used. When it isn't passed (the default), the following
name and email will be used for the git author:

   GIT_AUTHOR_NAME="invalid_git config"
   GIT_AUTHOR_EMAIL="<unknown@unknown>"

And a bare/incomplete header patch will be applied.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>