spp/scc: support preprocessor defines and "warn on missing files"
The scc/spp functionality that would warn on missing patch/kconf or
include files was incomplete. Since the flag (-w) was not implemented
in either scc or spp.

To implement this functionality, we make several changes:

 - add "-w" support to scc
 - support preprocessor defines in spp

With the ability to spport preprocessor defines, we can now annotate
scc files with:

 #define lazy_filenames

To enable the lazy initialization and validatation of filenames. This
means that missing files wil not be reported as errors until the tree
is processed, versus at preprocessing time.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>