kgit-s2q: always update ORIG_HEAD after applying changes
In situations where git am fails to apply patches, and git apply is used,
we must update ORIG_HEAD as well as HEAD. This is required, since if the
next patch in the queue also fails git am application, it will reset to
ORIG_HEAD before using git apply. If we haven't updated ORIG_HEAD, we'll
end up warping back to the top of the branch each time.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>